Adding jQuery and Twitter detection
Added a check for jQuery (the only global I could find that was easily hide-able is the main jQuery library function) and Twitter (if you embed Twitter's widgets.js script).  This obviously doesn't have as much impact as the other libraries, but it looks like jQuery's global variable footprint is pretty small, anyway.